+++ mobilertc-android-studio/example2/src/main/java/us/zoom/sdkexample2/Constants.java
@@ -20,4 +20,19 @@ public interface Constants {
 	// TODO Change it to your exist meeting ID to start meeting
     String MEETING_ID = null;
 
+    /**
+     * We recommend that, you can generate jwttoken on your own server instead of hardcore in the code.
+     * We hardcore it here, just to run the demo.
+     *
+     * You can generate a jwttoken on the https://jwt.io/
+     * with this payload:
+     * {
+     *     "appKey": "string", // app key
+     *     "iat": long, // access token issue timestamp
+     *     "exp": long, // access token expire time
+     *     "tokenExp": long // token expire time
+     * }
+     */
+    public final static String SDK_JWTTOKEN = YOUR JWTTOKEN;
+
 }