+++ mobilertc-android-studio/example2/src/main/java/us/zoom/sdkexample2/MainActivity.java
@@ -8,6 +8,7 @@ import us.zoom.sdk.StartMeetingOptions;
 import us.zoom.sdk.StartMeetingParams4APIUser;
 import us.zoom.sdk.ZoomError;
 import us.zoom.sdk.ZoomSDK;
+import us.zoom.sdk.ZoomSDKInitParams;
 import us.zoom.sdk.ZoomSDKInitializeListener;
 import android.app.Activity;
 import android.content.Intent;
@@ -50,7 +51,10 @@ public class MainActivity extends Activity implements Constants, MeetingServiceL
 		ZoomSDK zoomSDK = ZoomSDK.getInstance();
 		
 		if(savedInstanceState == null) {
-			zoomSDK.initialize(this, SDK_KEY, SDK_SECRET, WEB_DOMAIN, this);
+			ZoomSDKInitParams initParams = new ZoomSDKInitParams();
+			initParams.jwtToken = SDK_JWTTOKEN;
+			initParams.domain = WEB_DOMAIN;
+			zoomSDK.initialize(this, this, initParams);
 		}
 		
 		if(zoomSDK.isInitialized()) {
@@ -224,5 +228,10 @@ public class MainActivity extends Activity implements Constants, MeetingServiceL
 		if(meetingStatus == MeetingStatus.MEETING_STATUS_IDLE || meetingStatus == MeetingStatus.MEETING_STATUS_FAILED) {
 			selectTab(TAB_WELCOME);
 		}
+	}
+
+	@Override
+	public void onZoomAuthIdentityExpired() {
+
 	}
 }