+++ mobilertc-android-studio/sample/src/main/java/us/zoom/sdksample/initsdk/AuthConstants.java
@@ -11,4 +11,19 @@ public interface AuthConstants {
 	// TODO Change it to your APP Secret
 	public final static String SDK_SECRET = "";
 
+	/**
+	 * We recommend that, you can generate jwttoken on your own server instead of hardcore in the code.
+	 * We hardcore it here, just to run the demo.
+	 *
+	 * You can generate a jwttoken on the https://jwt.io/
+	 * with this payload:
+	 * {
+	 *     "appKey": "string", // app key
+	 *     "iat": long, // access token issue timestamp
+	 *     "exp": long, // access token expire time
+	 *     "tokenExp": long // token expire time
+	 * }
+	 */
+	public final static String SDK_JWTTOKEN = YOUR JWTTOKEN;
+
 }